back/requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
black==21.5b0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2.2
django-cache-memoize==0.1.9
django-cors-headers==3.7.0
django-environ==0.4.5
Faker==8.1.3
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
iniconfig==1.1.1
mixer==7.1.2
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
promise==2.3
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.2.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
Rx==1.6.1
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.4
